+++ lua/aerial/highlight.lua
@@ -69,7 +69,8 @@ else
   end
 end
 
-M.create_highlight_groups = function()
+-- Create highlight groups on initial load
+(function()
   -- Use Normal colors for AerialNormal, while stripping bold/italic/etc
   local normal_defn = get_hl_by_name("Normal")
   -- The default text highlight
@@ -143,6 +144,6 @@ M.create_highlight_groups = function()
   link("AerialStructIcon", "Type")
   link("AerialTypeParameterIcon", "Identifier")
   link("AerialVariableIcon", "Identifier")
-end
+end)()
 
 return M
