+++ tests/minimal_init.lua
@@ -31,10 +31,7 @@ if master_nvim_ts then
 
   vim.api.nvim_create_user_command("RunTests", function(opts)
     local path = opts.fargs[1] or "tests"
-    require("plenary.test_harness").test_directory(
-      path,
-      { minimal_init = "./tests/minimal_init.lua" }
-    )
+    require("plenary.test_harness").test_directory(path, { init = "./tests/minimal_init.lua" })
   end, { nargs = "?" })
 else
   -- Use compiler that includes c++14 features by default
