+++ tests/test_util.lua
@@ -139,7 +139,6 @@ end
 
 M.reset_editor = function()
   require("aerial").setup({})
-  require("aerial").sync_load()
   vim.cmd.tabonly({ mods = { silent = true } })
   for i, winid in ipairs(vim.api.nvim_tabpage_list_wins(0)) do
     if i > 1 then
